Improve safety and training pipeline
- RoPE: cast cos/sin to input dtype in _rope_apply to prevent fp16 AMP dtype mismatch
- set_seed: also seed Python random module for full reproducibility
- Wire up validation split (test_split config) with validate() for proper early stopping
- Use val loss for best loss/early stopping instead of train loss (which always decreases)
- Skip redundant per-epoch checkpoint save when epochs==1 (saves ~100MB)
- Print total effective steps per epoch in Training Config
- split_dataset: accept configurable seed parameter

+++ models/transformer.py
@@ -39,7 +39,7 @@ class RotaryEmbedding(nn.Module):
         d = x.size(-1) // 2
         x1, x2 = x[..., :d], x[..., d:]
         rot = torch.cat((-x2, x1), dim=-1)
-        return x * cos + rot * sin
+        return x * cos.to(x.dtype) + rot * sin.to(x.dtype)
 
 
 class CausalSelfAttention(nn.Module):
